Add files via upload
This file contains my first try at STEVE. The only thing that works at the moment is getting sentence representations. This is done by getting word representations by feeding their characters in a GRU and getting the hidden vector. These vectors then get concatenated and are fed into another GRU that does the same thing at a word level. At the moment the word embeddings are trained separately, in an auto encoder configuration. The same is later done with the sentence embeddings.
What does not work is the 'TSN' layer. This should transform the input sentence representation to the answer's representation. It is unable to produce something that gets decoded to anything meaningful at the moment. 
I will be working on implementing attention mechanisms.